Tighter Folder View item delegates.
Summary:
This patch reduces margins inside the item delegate. The most
frequent piece of user feedback for Folder View is "Can you let
me change the grid size?", and the background behind the question
always turns out to be unhappiness with the large gutters between
icons and not being able to fit more icons on the desktop. A
casual comparison to other desktops (including past KDE versions)
indicates icon delegates indeed tend to be tighter. This follows
suit.

On my screen, the desktop now fits 9 instead of 7 rows of icons,
with similar gains on the horizontal axis.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4856