[Folder View] Turn heading into a "hyperlink" and reduce hit area
I often end up accidentally opening the folder in a file manager because especially
when FolderView is placed on a desktop, the hit area for the label spans the entire
width and a significantly taller area than the label's text height actually is.

There's also no visual cue that clicking the heading would do anything.
Turn it into a "hyperlink" by underlining it on hover and using the PointingHandCursor.

Differential Revision: https://phabricator.kde.org/D4959