open applet handles when the toolbox is open
Summary:
this makes a mini "edit mode" when the toolbox is open, opening
all the applet handles.
It is coherent with the panel and partly solves the unintuitiveness
of the press and hold for the handle

Test Plan:
open the toolbox, clicking on an empty area or on another window closes it
clicking anywhere on an applet handle executes the action and doesn't close it.
is possible to resize, move and close applets without closing any handle or
the toolbox.

Unfortunately there is a timer more in the toolbox that i tried really hard to kill
but i didn't manage to have the correct behavior otherwise (ideas welcome)

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4969