Hoist the actions overlay out of the frame loader.
Summary:
This fixes a race condition causing the actions overlay to sometimes
appear in the drag pixmap attached to the DND cursor, as grabToImage
is run against the frame loader.

Test Plan:
We used to have a problem with the actions overlay not
repositioning properly in popup Folder Views when switching them
between list and icon view mode, which partly shaped the old code
(the anchor state changes). I verified this didn't regress with the
new simplified code.

Reviewers: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5309