Force KAccess to run under XCB
Summary:
KAccess is responsible for:

visual system bell
stick keys/slow keys
starting orca

None of this will work in wayland as-is; and when they are done, will
probably be in kwin anyway.

But, we do still want this set up for xwayland applications.

This sets the platform, so that we have xwayland apps still working the
same. Making wayland apps work properly is a task for another day.

Primarily this stops the thing crashing on startup.

BUG:372597

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6035