[Folder View] Pad cellWidth/cellHeight with the extra space
The right/bottom corners of the desktop will have an empty area where
icons should be touching. Dropping an icon there will cause a scrollbar.
https://bugs.kde.org/show_bug.cgi?id=380424

When when in a containment, use a floating point cellWidth/cellHeight
based on the avaiable space left. Folder widgets will still use a fixed
size based on the icon.

Note: We no longer attempt to fit the last column/row of icons under
the panels if there's just barely space for it. So this might cause
a scrollbar for those users.

Differential Revision: https://phabricator.kde.org/D6188