kimpanel: Reduce the inputpanel flicker.
The dbus message may come in very frequently because every field is
updated with a separate message (preedit, aux, lookup table). Try to
batch the GUI update with a timer. Also, do the hide at the beginning
and do the show at the end of ui update to reduce flicker furthermore.