Update size hints less often.
Summary:
The size hints were based on the size of pagerGridItem, which is derived
from its contents. This made the size hints update as delegates were
added, but we know ahead of time how many there will be, and what their
size will be, so we can save some busy work.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7384