Compress size hint changes before releasing position and repositioning item.
Summary:
D4204 / d7621da69b9b introduced more elaborate handling of item size
hint changes, releasing and repositioning an item for each change to
the min/max hints on either axis. We have applets in which the x and
y hints are coupled and update in turn, to implement aspect-ratio-
preserving resize. This makes the value on an individual axis a
useless temporary, wreaking havoc when followed.

As a poor means to commit size hint changes atomically, a 100ms timer
is used to compress size hint changes. The 100ms interval matches
layoutTimer in main.qml.

Not very happy with this.

Together with D7384 this addresses:
BUG:382759

Reviewers: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7385