kimpanel: change highlight color
Summary:
Before:

Highlighted candidate uses theme.highlightColor for text, but this color should only be used in background color, not text color. This style is different from other KDE UI.

{F3870601}

After:

Highlighted candidate uses Plasma's svg as highlight. PlasmaComponents.Highlight is not used here because it is View-specific item and we do not want use the View but keep using Layout here because view does not work well with the pattern like "always making view size to fit the content".

{F3874097}

It is more similar to general KDE design.

Reviewers: #plasma, guoyunhe

Reviewed By: guoyunhe

Subscribers: xuetianweng, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7436