Don't special case file URLs when processing .desktop files
Summary:
At first glance this looks faster, as we skip KIO for a simple stat(),
but as the result isn't cached we do file lookups on every single mouse
move.

This also has the side effect of not blocking Plasma when stat-ing a
file on a disconnected remote drive.

CCBUG: 379994

Test Plan:
Created two .desktop files one pointing to a dir, another to a file.
One had the arrow to expand the folder, the other did not.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7535