[Folder View Icons Config] Fix initial index of ComboBox
A while ago I optimized Plasma config dialogs to create the component with the cfg_ properties
already set. This code relied on the fact that mode changes initially.

Use Component.onCompleted now that we can and also use onActivated which only fires if the
index is explicitly changed by the user, not on imperative assignment.

Differential Revision: https://phabricator.kde.org/D7733