Sync XRDB DPI to the platform specific setting
Summary:
Even though this sets up the X rdb, we want to use the value the
user has set to use when under wayland - as X apps will be scaled by the
compositor.

This ensures when I open firefox on my X or Wayland session it's the
same size.

Colors KCM part is needed as it includes ../krdb/krdb.cpp

Test Plan:
Logged in on X with my normal settings, opened Firefox
Logged in on Wayland with my normal settings, opened Firefox

Same size on both

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7834