Add edit button to desktop theme
Summary:
This option is only visible in plasmathemeexplorer is installed. This
makes it easy to edit themes, without cluttering the UI of normal users.

Edit button is made similar to the action buttons in the wallpaper
delegate. (though this is copying an existing bug of of using
plasmacomponents where they shouldn't be)

MouseArea got collapsed into the parent Item in the delegate as it makes
for more readable implicit z-ordering to have the button on top.

Test Plan:
Clicked on one, explorer for that theme appeared.
Faked not having theme explorer by temporarily putting a typo in the service name. No buttons showed

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: jensreuterberg, apol, abetts, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8006