FolderView: position files at drop event target position
Summary:
Remember the drop event target position and move the items once
they become available. This requires us to allow moving while no
drag image is being shown. Otherwise, we listen to the
copy job signals, which gets created by the drop job. This allows
us to map target filenames in their final form, i.e. after the user
potentially renamed the files to handle conflicts, to some desired
visual position. To stay on the safe side, we also periodically
cleanup the mapping after an idle timeout of 10s. This ensures we
don't grow the mapping with stale items. This is required to handle
errors in the file lister, or situations like overwriting an existing
file which would not trigger a rowAdded signal.

Reviewers: hein, amantia

Reviewed By: hein, amantia

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8598