Update the proxy to source and source to proxy mapping correctly
Summary:
The problem is visible when an item is moved to another screen and moved back.
When it is moved out, the proxy-source mapping changes (as the items
is filtered out from the underlining model), while when moved back, it is
changed again, but in a wrong way that breaks the mapping and results
in "losing" one mapping, thus one item disappears.
There is a unit test added for it.

Reviewers: #plasma, hein, mwolff

Reviewed By: #plasma, hein, mwolff

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8915