Don't move back items to second screen if they were manually moved
Summary:
When a secondary screen is removed, the items from there are moved to
the first screen. If the screen is plugged back, the items are moved back.
This patch does not move back the items on screen reconnect if meantime they were manually
moved around on the first screen.

Reviewers: #plasma, hein, mwolff

Reviewed By: #plasma, hein

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9313