Slightly reduce vertical gutter width.
Removed 1 smallSpacing from the top and bottom padding of each cell.

Allows more rows to fit on screen. Results in less goofy-looking
vertical gutters with the default of two text rows (as most labels
only have one, it's a lot of dead space on screen in the common case).

Based on user feedback.

Reviewers: #plasma, hein

Differential Revision: https://phabricator.kde.org/D9201