Support font/ttf and font/otf mimetypes in kfontinst
Summary:
Those mimetypes are new and kfontinst/kfontview/fontthumbnail weren't aware
of that.
As the old types are aliases of the new types, it doesn't need to be added
to the file dialog options (it would create duplicate entries).
However, it broke thumbnail generation.

Test Plan: .ttf files have thumbnails again.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: anthonyfieroni, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9738