Fix launching by touchscreen tap
Summary:
The mouse handling code relied on onPositionChanged to do setup work
prior to onPressed. This is never guaranteed, but it's is likely to
happen when using the mouse (since the pointer is usually moved from
a panel icon to a delegate elsewhere on screen). In a touchscreen
scenario, these odds reverse.

The lesson to take away: Never assume onPositionChanged has run
before onPressed.

BUG:366527

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10513