missing return statement in trash applet
Summary:
The missing return was causing to set
Layout.minimumHeight in not correct value.
The issue was discovered when the trash
applet wasnt working with Latte parabolic
effect when it was in vertical docks.

Test Plan:
I made the change  and retested that everything works ok
in the panel or in the desktop.

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10979