fix detecting changes in sub-pixel and hinting settings
Summary:
Since D8916, on changes to the sub-pixel and hinting comboboxes the "apply" button stays disabled.
This patch fixes this, so that one can actually apply changes to the font rendering settings.

Test Plan: Open the fonts kcm and change the sub-pixel or hinting settings. The apply button should be enabled.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11174