[Mouse KCM] Add KWin Wayland backend
Summary:
The primal goal of this patch is the addition of a new backend in order to
configure pointer devices through KWin's libinput backend in a Wayland session.

The implementation builds upon Xuetian's backend splitting work, which again
was based on the backend splitting in the Touchpad KCM.

To integrate this backend nicely the code structure is cleaned up and Ui code
as well as backend specific code is pushed into the respective subdirectories
similar to the Touchpad KCM code.

{F5757933}

Test Plan: Tested the old backend on X and the new libinput backend on Wayland.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, ngraham, plasma-devel

Tags: #plasma

Maniphest Tasks: T7964

Differential Revision: https://phabricator.kde.org/D11468