Fix use of assert/crash when removing global shortcut
Summary:
I'm not the author, but it's pretty clear what the intention was.

We expect to get one result back from proxyModel->match
We use results.first(), so we want to check we have at least one entry.
The assert was written backwards

BUG: 387444

Test Plan: Can now remove a group without a crash

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11567