[Mouse KCM] Add X11 libinput exclusive backend and UI
Summary:
This patch splits up the current X11 backend into one for systems with
installed X libinput driver and systems with only evdev driver.

The evdev backend is used together with the old QWidget based UI. The
libinput backend is based on the KWin Wayland one and controlled by a
very similar QML based UI. One difference is that values are always
propagated to all pointer like devices and can not be set for
individual devices.

As for the evdev backend values are saved to a config file in the user
directory and reapplied on every session start.

Note that the libinput backend always takes precedence to the evdev one.
If an user wants to force the evdev backend, the X libinput driver needs
to be removed.

Depends on D11468

{F5757944}

BUG: 350688

Test Plan: Tested X session with and without libinput driver.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, hein, mart, abetts, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11469