Kicker: Make menus grow (to a limit) if the text doesn't fit on the default width
Summary:
On my install i have various application names that don't fit on units.gridUnit * 14
so let the list grow if needed. It only grows to a limit, we don't want the list to
be huge for some weirdly long application name.

Test Plan:
Use the menu editor to rename Development to Development 22222222222222222222222222
Check the root menu shows all the name
Search and check the search bar sizing is correct at all times
Open submenus with applications whose name previously was cut and check now it displays correctly

Note: I get
	TextFieldStyle.qml:46:17: QML QQuickItem: Binding loop detected for property "implicitHeight"
when entering text on the search bar, but i can also reproduce this with a simple testcase like
https://paste.kde.org/pi1hopivh so it seems it's not a bug in my code. Also i can't see the search
bar misbehaving so it seems to be "harmless"

Reviewers: #plasma, hein

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12171