add preview images to fonts kcm
Summary:
The fonts kcm offers different font rendering settings. Currently, one needs to apply the settings and
reopen the application to see the changes.
This patch adds a way to render fonts at different settings (mostly based on existing code in kfontinst) and
adds preview images to the sub-pixel and hinting comboboxes.

This is part of a planned redesign of the fonts kcm. See the discussion in https://phabricator.kde.org/T7927

Test Plan:
Open the fonts kcm, click on the sub-pixel combobox. The preview images should look different.

{F5743556}

Reviewers: #plasma, harmathy, mart, davidedmundson, ngraham

Reviewed By: #plasma, harmathy, mart, ngraham

Subscribers: ngraham, davidedmundson, abetts, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11064