Don't disable apply button when window is resized
Summary:
Before this patch if I edit an option, and then resize the window the
apply button became disabled.

I'm not sure what this was trying to do, but widget size and whether
we have changes to save seems very unrelated.

Test Plan: Tested as above, no longer have the bug

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12943