Fonts KCM: Fix text readability regression
Summary:
Porting the Fonts KCM to QML in 24b960f92284 set the state of the font
preview labels to `disabled`, resulting in a light gray text with too
little contrast to enable a usable preview of the chosen font.

In the `QWidgets` based KCM the following code was used:
  QLabel::setFrameStyle(QFrame::StyledPanel | QFrame::Sunken)
As that property is not available from QML, we use an alternative
approach by setting the state to `readOnly`. To also indicate
visually  that nothing can be typed into the textedit, we add
`Kirigami.Theme.inherit: true` which results in changing the
background from white to grey.

Test Plan:
Font previews for each category in `kcmshell5 kcm_fonts` should be
readable again.

Before: {F5894425}
After: {F5896658}
How it looked like in Plasma 5.12: {F5894423}

Reviewers: mart

Reviewed By: mart

Subscribers: mart, broulik, Zren, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13390