Touchpad KDED module: Convert to JSON metadata
The desktop file kded_touchpad.desktop sets X-KDE-DBus-ModuleName=touchpad,
which is presumably intended to register the KDED module under that
name on DBus.  However, this key appears to be being ignored, and the
module is registered under the default path "/modules/kded_touchpad".
This incorrect path means that touchpad control via the Plasma applet
does not work, and 'kcmshell5 kcmkded' does not show the touchpad module
status correctly.

Differential Revision: https://phabricator.kde.org/D13360