[Device Automounter] Load kded module only if enabled
There's no point in loading the entire module when automounting is disabled.
It does Solid device queries which can be quite expensive, albeit it delays initialization.

The AutomounterSettings::shouldAutomountDevice function allows force-mounting even if automounting as a whole is disabled
whereas the KCM UI does not. Not sure what the reason for that is.

Differential Revision: https://phabricator.kde.org/D13852