Improve arrow key navigation of Kicker search results
Summary:
Kicker already makes sure the first search result is highlighted
and actionable while searching. This patch makes arrow-down after
typing move to the second search result instead of merely moving
focus from the search field to the list, so the user doesn't have
to press arrow-down twice anymore.

It also allows using arrow-left/right to jump to the other result
columns, when the cursor is at the start or end of the text field,
respectively.

BUG:397779

Reviewers: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15286