[Folder View] Use KIO::PreviewJob::defaultPlugins() for default plugins
Rather than a hardcoded list of only two thumbnailers.
All of this is done under the hood so the QML part doesn't need to be changed.
This is the same behavior as Dolphin uses, when the list of thumbnailers is empty, the default set (which includes all thumbnailers
except textthumbs) is used.

CHANGELOG: Folder View now enables previews for most file types by default

Differential Revision: https://phabricator.kde.org/D15268