[Folder View] Remember selected item when navigating in subfolders
Summary:
When using the folder view in a panel as a popup, being in list view
mode, getting into subfolders and back with arrow keys had a weird
behaviour: the currentItem would just stay the same accross folders.

This patch restores both the proper selected item and the scrolling
position when navigating back.

BUG: 359310

Test Plan:
- Add a folder view in a panel, configure it in list view mode
- Enter a subfolder and go back
The folder you had gone into should be selected and the scrolling
position should be the same than when you entered it.

Reviewers: #plasma, broulik, hein

Reviewed By: #plasma, hein

Subscribers: anthonyfieroni, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15840