[Kickoff] Add a subtle separator line between the header and the content view
Summary: This is another idea that shook out of D15011: add subtle lines to separate the main content area from both the tab bar and the header, giving Kickoff some much-needed structure while still not overwhelming the user with lines and frames and chrome.

Test Plan:
Tested with Kickoff on a panel on all four different sides of the screen, and with Breeze light and Breeze dark themes.

Examples with the default panel:
{F6348699}
{F6347337}

Example with the panel on the right side: {F6347340}

Reviewers: #plasma, #vdg

Subscribers: abetts, mmustac, hein, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15206