[Folder View] implement a minimum width for icon view to ensure that labels are never rendered useless
Summary:
Currently with Folder-View-on-the-desktop, items' labels become illegibe when the icon size is very small.

This patch implements a minimum width when using on-the-desktop icon view to make sure that the label is never uselessly short when using very small icons.

BUG: 379432
FIXED-IN: 5.12.8

Test Plan:
Smallest size, before: {F6426124}
Smallest size, after: {F6426126}
Equal padding on either side of the container itself still works: {F6426129}

There is no change for sizes corresponding to the fourth tick on the slider and beyond

Reviewers: #plasma, #vdg, hein

Reviewed By: #plasma, hein

Subscribers: abetts, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D16901