[Componentchooser KCM] Make default browser app choice work even when combobox is not used
Summary:
BUG: 350663
FIXED-IN: 5.12.8

When choosing a default browser app using the app list combobox, the choice was only being saved after the combobox was used, because its signal was connected to an inline function that saved the choice, while the radio button that enabled the combobox was not. As a result, if you clicked on that radio button and didn't need to use the combobox because it was already showing your preferred browser, that action was not actually getting saved.

This patch fixes the situation my moving the inline function into a real explicit slot function and connecting the radio button to it as well.

Test Plan:
- Go to the component chooser KCM's default browser module
- Reset to defaults, then leave the KCM and open it again
- Click the {nav in the following application} radio button
- Do NOT touch the combobox
- Click {nav Apply}
- Leave the KCM and come back

With this patch, the radio button's state is preserved and the correct browser is selected in the combobox.

Reviewers: #plasma, cfeck, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17181