Add an X-DocPath link to the Activities KCM
Summary:
Even if the documentation does not explain the KCM in details,
at least it provides a basic introduction to the concept.
Moreover, the Help button is currently enabled but it does not
lead to a valid help page, so something should be fixed anyway.

Test Plan:
The Help button of the Activities KCM opens an help page
about activities.

Reviewers: davidedmundson, ngraham

Reviewed By: davidedmundson, ngraham

Subscribers: ngraham, kde-doc-english, plasma-devel

Tags: #plasma, #documentation

Differential Revision: https://phabricator.kde.org/D18145