[kcms/cursortheme] Fix size combo
Summary:
The biggest problem fixed was that m_preferredSize was sometimes treated
as
an absolute value, sometimes as an index. This patch splits the two
concepts.

It also fixes the binding of the combobox to handle updates properly.

BUG: 403019

Test Plan:
Clicked between themes
Closed and reopened the KCM had a valid value

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18471