[Desktop Toolbox] Use a clearer name when the label is shown
Summary:
When the Desktop Toolbox is dragged out of a corner, its label is shown. Currently this label is not very clear since it consists only of the name of the current activity, which by default is "Default".

This patch introduces three changes:
- When the label is shown, always prepend it with the widget's actual name: "Desktop Toolbox"
- Append the word "Activity" so it's obvious what the preceding string refers to
- Don't show the activity name if it's "Default" since this signifies that the user does not use Activities and would be confused by seeing references to a "Default Activity".

Test Plan:
There are no visible changes to the default appearance; the Desktop Toolbox still has no label when it's in a corner. When dragged out of the corner, here's how it looks with this patch:

When using the Default activity: {F6567804}

When using differently-named activity: {F6567805}

Reviewers: #vdg, #plasma, hein, GB_2

Reviewed By: #vdg, #plasma, hein, GB_2

Subscribers: GB_2, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18528