Disable KCMs while downloading file
When dropping or choosing a remote file, it is downloaded to a temporary location for installation.
Disable the KCM while this is in progress to avoid confusion as to what is happening.
A job is spawned in notification area which is why no additional busy indicator is added to the KCM UI.

Differential Revision: https://phabricator.kde.org/D18497