[kcms/keys] Fix double margin
Summary:
A widget should expand to the edge of the widget. i.e a QLineEdit covers
the size of the whole line edit. It's up to the parent layout to then
provide margins.

The same rule applies to complex widgets which are a layout of other
widgets.

Test Plan: Frame lines up with bottom buttons

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18753