[containment] Show an inline message when the Desktop Toolbox is hidden
Summary:
Currently, when the Desktop Toolbox is hidden, the only way to return to the
containment configuration window is by right-clicking on the desktop. This may
not be obvious or may be difficult for some people sue to age, disability,
hardware, etc.

This patch makes an informational message appear when the user goes to hide the
Desktop Toolbox that explains how to return to the configuration window in the
future. It only appears when the checkbox is currently checked and it is unchecked;
it does not appear when the user visits the {nav Tweaks} page when the checkbox is
already unchecked (by then they already know because they've already seen the
message at least once)

Test Plan: {F6613733, size=full}

Reviewers: #vdg, #plasma, mart

Reviewed By: #vdg, #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18991