[Desktop Sessions KCM] Add Restart to BIOS/UEFI checkbox
This sets a logind flag that tells the system to boot into the BIOS/UEFI setup screen on next boot.
It can be quite a challenge to enter the setup screen on boot these days as with quick boot and what not
the timeframe for the keypress is often quite narrow.

Differential Revision: https://phabricator.kde.org/D19560