foldermodel: store information about screen used
Summary:
BUG: 401464

According to FolderModel::filterAcceptsRow code, it is possible that folderview is used by containment but not associated with a screen.
In case of m_screen == -1 there was no information about which items we should filter out.

Instead, we do not change m_screen to -1 and we can filter out items that does not belong to our screen.
m_screenUsed shows whether folderview is associated with a screen or not

Reviewers: broulik, hein, #plasma

Reviewed By: hein, #plasma

Subscribers: cfeck, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18979